+++ v3/crates/graphql/ir/src/lib.rs
@@ -29,7 +29,7 @@ pub use global_id::{global_id_col_format, GLOBAL_ID_VERSION};
 pub use model_selection::ModelSelection;
 pub use model_tracking::get_all_usage_counts_in_query;
 pub use mutation_root::generate_ir as generate_mutation_ir;
-pub use order_by::ResolvedOrderBy;
+pub use order_by::OrderBy;
 pub use permissions::process_model_predicate;
 pub use plan::{
     generate_request_plan, plan_expression, process_model_relationship_definition,