+++ v3/crates/graphql/ir/src/model_selection.rs
@@ -15,11 +15,7 @@ use serde::Serialize;
 use std::collections::BTreeMap;
 use std::sync::Arc;
 
-use super::{
-    aggregates, arguments, filter,
-    order_by::{self, ResolvedOrderBy},
-    permissions, selection_set,
-};
+use super::{aggregates, arguments, filter, order_by, permissions, selection_set};
 use crate::error;
 use crate::model_tracking::count_model;
 use graphql_schema::GDS;
@@ -48,7 +44,7 @@ pub struct ModelSelection<'s> {
     pub offset: Option<u32>,
 
     // Order by
-    pub order_by: Option<ResolvedOrderBy<'s>>,
+    pub order_by: Option<order_by::OrderBy<'s>>,
 
     // Fields requested from the model
     pub selection: Option<selection_set::ResultSelectionSet<'s>>,
@@ -65,7 +61,7 @@ struct ModelSelectAggregateArguments<'s> {
 struct FilterInputArguments<'s> {
     limit: Option<u32>,
     offset: Option<u32>,
-    order_by: Option<ResolvedOrderBy<'s>>,
+    order_by: Option<order_by::OrderBy<'s>>,
     filter_clause: Option<Expression<'s>>,
 }
 
@@ -80,7 +76,7 @@ pub fn model_selection_ir<'s>(
     permissions_predicate: &'s metadata_resolve::FilterPermission,
     limit: Option<u32>,
     offset: Option<u32>,
-    order_by: Option<ResolvedOrderBy<'s>>,
+    order_by: Option<order_by::OrderBy<'s>>,
     session_variables: &SessionVariables,
     request_headers: &reqwest::header::HeaderMap,
     usage_counts: &mut UsagesCounts,
@@ -322,6 +318,7 @@ fn read_filter_input_arguments<'s>(
                     }
                     order_by = Some(order_by::build_ndc_order_by(
                         filter_input_field_arg,
+                        session_variables,
                         usage_counts,
                     )?);
                 }
@@ -374,7 +371,7 @@ fn model_aggregate_selection_ir<'s>(
     permissions_predicate: &'s metadata_resolve::FilterPermission,
     limit: Option<u32>,
     offset: Option<u32>,
-    order_by: Option<ResolvedOrderBy<'s>>,
+    order_by: Option<order_by::OrderBy<'s>>,
     session_variables: &SessionVariables,
     usage_counts: &mut UsagesCounts,
 ) -> Result<ModelSelection<'s>, error::Error> {